Add missing Q_OBJECT macros
