+++ src/3d/qgsglobechunkedentity.h
@@ -34,9 +34,66 @@
 
 #include "qgschunkedentity.h"
 
+#include "qgschunkloader.h"
+#include "qgscoordinatetransform.h"
+#include "qgsdistancearea.h"
+#include <QImage>
+
 class QgsMapLayer;
 class QgsGlobeMapUpdateJobFactory;
+class QgsTerrainTextureGenerator;
 
+class QgsGlobeChunkLoader : public QgsChunkLoader
+{
+    Q_OBJECT
+  public:
+    QgsGlobeChunkLoader( QgsChunkNode *node, QgsTerrainTextureGenerator *textureGenerator, const QgsCoordinateTransform &globeCrsToLatLon );
+
+    Qt3DCore::QEntity *createEntity( Qt3DCore::QEntity *parent ) override;
+
+  private:
+    QgsTerrainTextureGenerator *mTextureGenerator;
+    QgsCoordinateTransform mGlobeCrsToLatLon;
+    int mJobId;
+    QImage mTexture;
+};
+
+
+//! Handles asynchronous updates of globe's map images when layers change
+class QgsGlobeMapUpdateJob : public QgsChunkQueueJob
+{
+    Q_OBJECT
+  public:
+    QgsGlobeMapUpdateJob( QgsTerrainTextureGenerator *textureGenerator, QgsChunkNode *node );
+
+    void cancel() override;
+
+  private:
+    QgsTerrainTextureGenerator *mTextureGenerator = nullptr;
+    int mJobId;
+};
+
+class QgsGlobeChunkLoaderFactory : public QgsChunkLoaderFactory
+{
+    Q_OBJECT
+  public:
+    QgsGlobeChunkLoaderFactory( Qgs3DMapSettings *mapSettings );
+
+    ~QgsGlobeChunkLoaderFactory();
+
+    QgsChunkLoader *createChunkLoader( QgsChunkNode *node ) const override;
+
+    QgsChunkNode *createRootNode() const override;
+
+    QVector<QgsChunkNode *> createChildren( QgsChunkNode *node ) const override;
+
+  private:
+    Qgs3DMapSettings *mMapSettings = nullptr;
+    QgsTerrainTextureGenerator *mTextureGenerator = nullptr; // owned by the factory
+    QgsDistanceArea mDistanceArea;
+    QgsCoordinateTransform mGlobeCrsToLatLon;
+    double mRadiusX, mRadiusY, mRadiusZ;
+};
 
 /**
  * 3D chunked entity implementation to generate globe mesh with constant elevation
